requirements.txt
langgraph>=0.2
langchain-core>=0.3
langchain-openai>=0.2
openai>=1.0
uvicorn>=0.15
python-dotenv>=1.0
fastapi>=0.68